+++ src/main/java/com/sajotuna/books/book/domain/Book.java
@@ -46,9 +46,6 @@ public class Book {
 
     private Double sellingPrice;
 
-    @Column(name = "stock") // 재고 필드 추가
-    private Integer stock;
-
     private Boolean giftWrappingAvailable;
 
     private Integer likes; // 좋아요 수
@@ -74,7 +71,7 @@ public class Book {
     // 생성자 (필요에 따라 추가)
     public Book(String isbn, String title, String author, String publisher, LocalDate publicationDate,
                 Integer pageCount, String imageUrl, String description,
-                Double originalPrice, Double sellingPrice, Integer stock, Boolean giftWrappingAvailable, Integer likes) {
+                Double originalPrice, Double sellingPrice, Boolean giftWrappingAvailable, Integer likes) {
         this.isbn = isbn;
         this.title = title;
         this.author = author;
@@ -85,7 +82,6 @@ public class Book {
         this.description = description;
         this.originalPrice = originalPrice;
         this.sellingPrice = sellingPrice;
-        this.stock = stock; // 생성자에 재고 필드 추가
         this.giftWrappingAvailable = giftWrappingAvailable;
         this.likes = likes;
         this.viewCount = 0;
@@ -136,7 +132,6 @@ public class Book {
         this.description = request.getDescription();
         this.originalPrice = request.getOriginalPrice();
         this.sellingPrice = request.getSellingPrice();
-        this.stock = request.getStock(); // 재고 업데이트 추가
         this.giftWrappingAvailable = request.getGiftWrappingAvailable();
         // 좋아요 수는 업데이트 시 변경하지 않음 (따로 관리)
     }
@@ -144,10 +139,5 @@ public class Book {
     // 좋아요 수 업데이트 메서드 추가 (관리자 기능)
     public void updateLikes(Integer newLikes) {
         this.likes = newLikes;
-    }
-
-    // 재고 수 업데이트 메서드 추가 (관리자 기능)
-    public void updateStock(Integer newStock) {
-        this.stock = newStock;
     }
 }
