+++ src/main/resources/application-prod.yml
@@ -1,0 +1,8 @@
+spring:
+  application:
+    name: book-api
+  config:
+    import: configserver:http://localhost:10379
+  cloud:
+    config:
+      label: dev
