+++ src/main/resources/application.yml
@@ -1,59 +1,5 @@
 spring:
-  datasource:
-    url: jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
-    driver-class-name: org.h2.Driver
-    username: sa
-    password:
-
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
-
   application:
     name: book-api
-
-  jpa:
-    hibernate:
-      ddl-auto: create
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-
-  elasticsearch:
-    uris: s4.java21.net:9200
-    username: elastic
-    password: nhnacademy123!
-
-
-server:
-  port: ${PORT:10364}
-
-#eureka:
-#  client:
-#    service-url:
-#      defaultZone: http://localhost:10372/eureka/
-#    healthcheck:
-#      enabled: true
-#  instance:
-#    prefer-ip-address: true
-#    instance-id: ${spring.application.name}:${server.port}
-#
-#management:
-#  endpoints:
-#    web:
-#      exposure:
-#        include: health,info,shutdown
-
-management:
-  health:
-    elasticsearch:
-      enabled: false
-
-
-logging:
-  level:
-    com.netflix.discovery: off
-    com.netflix.eureka: off
-    org.springframework.cloud.netflix.eureka: off
+  profiles:
+    active: prod
